requirements.txt
# Core LlamaIndex
llama-index>=0.11.0

# LlamaIndex Integrations
llama-index-readers-google>=0.3.0
llama-index-vector-stores-pinecone>=0.1.2
llama-index-embeddings-openai>=0.2.0
llama-index-llms-openrouter>=0.2.0

# Vector Database
pinecone-client>=3.0.0

# Frontend
streamlit>=1.30.0

# Environment Management
python-dotenv>=1.0.0

# Google Auth (required by GoogleDriveReader)
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0
